+++ src/mpi/coll/src/coll_impl.c
@@ -5,6 +5,7 @@
 
 #include "mpiimpl.h"
 #include "coll_impl.h"
+#include "coll_csel.h"
 
 /*
 === BEGIN_MPI_T_CVAR_INFO_BLOCK ===
